fix: refactor label filter parser and renderer into separate reusable functions
