+++ packages/ui/src/views/repo/pull-request/pull-request-list-page.tsx
@@ -26,7 +26,7 @@ import { getPRListFilterOptions, getSortDirections, getSortOptions } from '../co
 import { useFilters } from '../hooks'
 import { filterPullRequests } from '../utils/filtering/pulls'
 import { sortPullRequests } from '../utils/sorting/pulls'
-import { LabelsFilter, LabelsValue } from './components/labels'
+import { filterLabelRenderer, getParserConfig, LabelsFilter, LabelsValue } from './components/labels'
 import { PullRequestList as PullRequestListContent } from './components/pull-request-list'
 import type { PRListFilters, PullRequestPageProps } from './pull-request.types'
 
@@ -70,29 +70,7 @@ const PullRequestListPage: FC<PullRequestPageProps> = ({
     label: t('views:repos.prListFilterOptions.labels.label', 'Label'),
     value: 'label_by',
     type: FilterFieldTypes.Custom,
-    parser: {
-      parse: (value: string): LabelsValue => {
-        const result: LabelsValue = {}
-
-        value.split(';').forEach(entry => {
-          const [key, valueStr = ''] = entry.split(':')
-          if (!key) return
-          result[key] = valueStr === 'true' ? true : valueStr
-        })
-
-        return result
-      },
-      serialize: (value: LabelsValue): string => {
-        const parts = Object.entries(value)
-          .map(([key, val]) => {
-            if (!val) return ''
-            return `${key}:${val}`
-          })
-          .filter(Boolean)
-
-        return `${parts.join(';')}`
-      }
-    },
+    parser: getParserConfig(),
     filterFieldConfig: {
       renderCustomComponent: function ({ value, onChange }): ReactNode {
         return (
@@ -106,37 +84,12 @@ const PullRequestListPage: FC<PullRequestPageProps> = ({
           />
         )
       },
-      renderFilterLabel: (value?: LabelsValue) => {
-        const labelValuesArr = Object.entries(value ?? {}).filter(([_, val]) => val === true || val)
-        const [firstKey, firstValue] = labelValuesArr[0] || []
-        if (!firstValue) return ''
-
-        const labelDetails = labels.find(label => String(label.id) === firstKey)
-        const valueDetails = labelValueOptions[labelDetails?.key ?? '']?.find(value => String(value.id) === firstValue)
-
-        const remainingLabelValues = labelValuesArr.length - 1
-        return (
-          <div className="flex w-max gap-1">
-            {labelDetails && (
-              <LabelMarker
-                key={firstKey}
-                color={labelDetails.color}
-                label={labelDetails.key}
-                value={valueDetails?.value || ''}
-              />
-            )}
-
-            {remainingLabelValues > 0 && (
-              <Badge
-                size="sm"
-                className="text-xs"
-                variant="counter"
-                theme="primary"
-              >{`+ ${remainingLabelValues}`}</Badge>
-            )}
-          </div>
-        )
-      }
+      renderFilterLabel: (value?: LabelsValue) =>
+        filterLabelRenderer({
+          selectedValue: value,
+          labelOptions: labels,
+          valueOptions: labelValueOptions
+        })
     }
   }
